Refactor to use new class

+++ run.sh
@@ -6,11 +6,20 @@ export PORT=3978
 lsof -i :$PORT
 kill -9 $(lsof -sTCP:LISTEN -i:$PORT -t)
 
-echo "Removing node modules folder and installing latest"
-rm -rf node_modules
-rm package-lock.json
+if [ $1 == "clean" ]
+then
+    echo "Remove node modules folder and package-lock"
+    rm -rf node_modules
+    rm package-lock.json
+fi
+
+echo "Check for module updates"
 ncu -u
+
+echo "Install updates"
 npm install
+
+echo "Check for security issues"
 npm audit fix
 snyk test
